chaos: rule types and effect-selector validation (#4348)
* intproxy session-monitor: add chaos rule and request types
Add types for rules themselves as well as requests, so we can parse JSON into
a request type and then validate that the rule is logically sound. May see like
overkill but getting rid of the JSON early is much nicer for validation.

* chaos rules: impl TryFrom value for chaos request type
Badly formed requests will fail here, for example multiple effects
in one rule, missing selectors etc.

* chaos rules: add tests and docs
Add tests for valid rules, valid requests for invalid rules and invalid
requests. Some of this can be refactored into test fixtures to make the
code easier to read.

* chaos rules: add empty post-mvp types, return readable errors from validation and parsing

Also switched to using `skip_serializing_none` from serde_with

* meta: move chaos rule and request types to rules.rs

* merge: merge cleanup
