+++ Cargo.toml
@@ -205,8 +205,8 @@ serde_json = "1"
 serde_json_path = "0.7.2"
 serde_json_path_macros = "0.1.6"
 serde_urlencoded = "0.7"
-serde_yaml = "0.9"
 serde_with = "3.18"
+serde_yaml = "0.9"
 sha2 = "0.10"
 shellexpand = "3"
 socket2 = { version = "0.5", features = ["all"] }
