+++ mirrord/intproxy/Cargo.toml
@@ -31,7 +31,8 @@ mirrord-protocol-io = { path = "../protocol-io" }
 
 futures.workspace = true
 semver.workspace = true
-serde = { workspace = true }
+serde.workspace = true
+serde_with.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
@@ -47,14 +48,12 @@ strum_macros.workspace = true
 tokio-retry.workspace = true
 tokio-rustls.workspace = true
 tokio-util.workspace = true
+uuid.workspace = true
+anyhow.workspace = true
 
 axum.workspace = true
 serde_json.workspace = true
 mirrord-session-monitor-protocol.workspace = true
-
-uuid = { workspace = true }
-anyhow = { workspace = true }
-serde_with = { workspace = true }
 
 [target.'cfg(windows)'.dependencies]
 winapi = { workspace = true }
